Add Task System

+++ app/src/main/java/com/yistudio/capital/engine/event/GameEvent.kt
@@ -2,7 +2,7 @@ package com.yistudio.capital.engine.event
 
 sealed class GameEvent {
     // 资产变动事件（用于触发资产规模任务）
-    data class CapitalChanged(val totalCapital: Long) : GameEvent()
+    data class CapitalChanged(val totalCapital: Double) : GameEvent()
 
     // 业务升级事件（用于触发等级任务）
     data class BusinessUpgraded(val businessId: String, val newLevel: Int) : GameEvent()
@@ -11,5 +11,5 @@ sealed class GameEvent {
     object ManualClick : GameEvent()
 
     // 收益变动事件（用于触发秒收益任务）
-    data class IncomeChanged(val incomePerSecond: Long) : GameEvent()
+    data class IncomeChanged(val incomePerSecond: Double) : GameEvent()
 }
